перенёс поля и методы из SaverOutput в MyWriter
