+++ sementsov aleksandr/BioinformStrong-SIGN/com/company/MyWriter.java
